NBA_DataLake/requirement.txt
boto3==1.34.28
requests==2.31.0
python-dotenv==1.0.0
